move code to src; extend Makefile; add new links for zip file; add zip for period 19; change all 19_20 scripts to 20 only; add tqdm and remove custom progress bar
