+++ python/pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "regex>=2024.5.15",
     "requests>=2.32.3",
     "sqlalchemy>=2.0.31",
+    "tqdm>=4.62.3"
 ]
 
 [project.optional-dependencies]
