+++ python/src/open_discourse/02_preprocessing/04_extract_mps_from_mp_base_data.py
@@ -113,4 +113,5 @@ mps["constituency"] = mps["constituency"].str.replace("[)(]", "", regex=True)
 mps = mps.astype(dtype={"ui": "int64", "birth_date": "str", "death_date": "str"})
 
 mps.to_pickle(save_path)
-print("Done.")
+
+print("Script 02_07 done.")
