+++ python/src/open_discourse/04_politicians/03_merge_politicians.py
@@ -2,12 +2,13 @@ import pandas as pd
 import regex
 
 import open_discourse.definitions.path_definitions as path_definitions
-from open_discourse.helper_functions.progressbar import progressbar
+from tqdm import tqdm
 
 # input directory
 MGS_PATH = path_definitions.POLITICIANS_STAGE_01
 MPS_PATH = path_definitions.POLITICIANS_STAGE_02
 FACTIONS_PATH = path_definitions.DATA_FINAL
+
 mps = pd.read_pickle(MPS_PATH / "mps.pkl")
 mgs = pd.read_pickle(MGS_PATH / "mgs.pkl")
 factions = pd.read_pickle(FACTIONS_PATH / "factions.pkl")
@@ -140,7 +141,7 @@ for (
     position_from,
     position_until,
     faction,
-) in progressbar(
+) in tqdm(
     zip(
         mgs["last_name"],
         mgs["first_name"],
@@ -151,7 +152,7 @@ for (
         mgs["position_until"],
         mgs["faction"],
     ),
-    "Merging mp-data...",
+    desc="Merging mp-data...",
 ):
     # Hardcode special cases
     if last_name == "Fischer" and first_name[0] == "Joschka":
@@ -281,4 +282,7 @@ for (
                 }
                 series = pd.DataFrame(series, index=[politicians.index[-1]])
                 politicians = pd.concat([politicians, series], ignore_index=True)
+
 politicians.to_csv(FACTIONS_PATH / "politicians.csv", index=False)
+
+print("Script 04_03 done.")
